implement support for *os.Root
fixes https://github.com/google/renameio/issues/44

+++ tempfile.go
@@ -56,6 +56,33 @@ func openTempFile(dir, name string, perm os.FileMode) (*os.File, error) {
 	}
 }
 
+// openTempFileRoot creates a randomly named file in root and returns an open
+// handle. It is similar to os.CreateTemp except that the directory must be
+// given, the file permissions can be controlled and patterns in the name are
+// not supported.  The name is always suffixed with a random number.
+func openTempFileRoot(root *os.Root, name string, perm os.FileMode) (string, *os.File, error) {
+	prefix := name
+
+	for attempt := 0; ; {
+		// Generate a reasonably random name which is unlikely to already
+		// exist. O_EXCL ensures that existing files generate an error.
+		name := prefix + strconv.FormatInt(nextrandom(), 10)
+
+		f, err := root.OpenFile(name, os.O_RDWR|os.O_CREATE|os.O_EXCL, perm)
+		if !os.IsExist(err) {
+			return name, f, err
+		}
+
+		if attempt++; attempt > 10000 {
+			return "", nil, &os.PathError{
+				Op:   "tempfile",
+				Path: name,
+				Err:  os.ErrExist,
+			}
+		}
+	}
+}
+
 // TempDir checks whether os.TempDir() can be used as a temporary directory for
 // later atomically replacing files within dest. If no (os.TempDir() resides on
 // a different mount point), dest is returned.
@@ -116,6 +143,8 @@ type PendingFile struct {
 	done           bool
 	closed         bool
 	replaceOnClose bool
+	root           *os.Root
+	tmpname        string
 }
 
 // Cleanup is a no-op if CloseAtomicallyReplace succeeded, and otherwise closes
@@ -132,8 +161,14 @@ func (t *PendingFile) Cleanup() error {
 	if !t.closed {
 		closeErr = t.File.Close()
 	}
-	if err := os.Remove(t.Name()); err != nil {
-		return err
+	if t.root != nil {
+		if err := t.root.Remove(t.tmpname); err != nil {
+			return err
+		}
+	} else {
+		if err := os.Remove(t.Name()); err != nil {
+			return err
+		}
 	}
 	t.done = true
 	return closeErr
@@ -161,8 +196,14 @@ func (t *PendingFile) CloseAtomicallyReplace() error {
 	if err := t.File.Close(); err != nil {
 		return err
 	}
-	if err := os.Rename(t.Name(), t.path); err != nil {
-		return err
+	if t.root != nil {
+		if err := t.root.Rename(t.tmpname, t.path); err != nil {
+			return err
+		}
+	} else {
+		if err := os.Rename(t.Name(), t.path); err != nil {
+			return err
+		}
 	}
 	t.done = true
 	return nil
@@ -198,6 +239,7 @@ type config struct {
 	ignoreUmask     bool
 	chmod           *os.FileMode
 	renameOnClose   bool
+	root            *os.Root
 }
 
 // NewPendingFile creates a temporary file destined to atomically creating or
@@ -225,8 +267,15 @@ func NewPendingFile(path string, opts ...Option) (*PendingFile, error) {
 	}
 
 	if cfg.attemptPermCopy {
+		var existing os.FileInfo
+		var err error
+		if cfg.root != nil {
+			existing, err = cfg.root.Lstat(cfg.path)
+		} else {
+			existing, err = os.Lstat(cfg.path)
+		}
 		// Try to determine permissions from an existing file.
-		if existing, err := os.Lstat(cfg.path); err == nil && existing.Mode().IsRegular() {
+		if err == nil && existing.Mode().IsRegular() {
 			perm := existing.Mode() & os.ModePerm
 			cfg.chmod = &perm
 
@@ -238,7 +287,14 @@ func NewPendingFile(path string, opts ...Option) (*PendingFile, error) {
 		}
 	}
 
-	f, err := openTempFile(tempDir(cfg.dir, cfg.path), "."+filepath.Base(cfg.path), cfg.createPerm)
+	var f *os.File
+	var err error
+	var tmpname string
+	if cfg.root != nil {
+		tmpname, f, err = openTempFileRoot(cfg.root, "."+filepath.Base(cfg.path), cfg.createPerm)
+	} else {
+		f, err = openTempFile(tempDir(cfg.dir, cfg.path), "."+filepath.Base(cfg.path), cfg.createPerm)
+	}
 	if err != nil {
 		return nil, err
 	}
@@ -253,7 +309,13 @@ func NewPendingFile(path string, opts ...Option) (*PendingFile, error) {
 		}
 	}
 
-	return &PendingFile{File: f, path: cfg.path, replaceOnClose: cfg.renameOnClose}, nil
+	return &PendingFile{
+		File:           f,
+		path:           cfg.path,
+		replaceOnClose: cfg.renameOnClose,
+		root:           cfg.root,
+		tmpname:        tmpname,
+	}, nil
 }
 
 // Symlink wraps os.Symlink, replacing an existing symlink with the same name
@@ -284,6 +346,44 @@ func Symlink(oldname, newname string) error {
 	}
 
 	if err := os.Rename(symlink, newname); err != nil {
+		return err
+	}
+
+	cleanup = false
+	return os.RemoveAll(d)
+}
+
+// SymlinkRoot wraps os.Symlink, replacing an existing symlink with the same
+// name atomically (os.Symlink fails when newname already exists, at least on
+// Linux).
+func SymlinkRoot(root *os.Root, oldname, newname string) error {
+	// Fast path: if newname does not exist yet, we can skip the whole dance
+	// below.
+	if err := root.Symlink(oldname, newname); err == nil || !os.IsExist(err) {
+		return err
+	}
+
+	// We need to use os.MkdirTemp, as we cannot overwrite a os.CreateTemp file,
+	// and removing+symlinking creates a TOCTOU race.
+	//
+	// There is no os.Root-compatible os.MkdirTemp, so we use the path directly.
+	d, err := os.MkdirTemp(root.Name(), "."+filepath.Base(newname))
+	if err != nil {
+		return err
+	}
+	cleanup := true
+	defer func() {
+		if cleanup {
+			os.RemoveAll(d)
+		}
+	}()
+
+	symlink := filepath.Join(filepath.Base(d), "tmp.symlink")
+	if err := root.Symlink(oldname, symlink); err != nil {
+		return err
+	}
+
+	if err := root.Rename(symlink, newname); err != nil {
 		return err
 	}
